Refactor: Move SuperAdminEmailContent to super-admin/email folder
